add receipt crud

+++ GroupExpenses.BLL/ViewModels/ReceiptViewModel.cs
@@ -7,8 +7,8 @@ namespace GroupExpenses.BLL.ViewModels
       public int Id { get; set; }
       public int EventId { get; set; }
       public string Name { get; set; }
-      public double Price { get; set; }
-      public double PriceInEur { get; set; }
+      public decimal Price { get; set; }
+      public decimal PriceInEur { get; set; }
       public string Description { get; set; }
       public Currency Currency { get; set; }
       public UserViewModel PaidBy { get; set; }
